Crate pallet_prices

Prices pallet

Overview

This pallet provides the price from Oracle Module by implementing the PriceFeeder trait. In case of emergency, the price can be set directly by Oracle Collective.

Re-exports

pub use pallet::*;

Modules

pallet
The module that hosts all the FRAME types needed to add this pallet to a runtime.