Module pallet_prices::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

Pallet
The pallet implementing the on-chain logic.

_GeneratedPrefixForStorageEmergencyPrice

Enums

Call
Contains one variant per dispatchable that can be called by an extrinsic.

Event
The event emitted by this pallet.

Traits

Config
Configuration trait of this pallet.

Type Definitions

EmergencyPrice
Mapping from currency id to it’s emergency price

Module	Deprecated
Type alias to Pallet, to be used by construct_runtime.