Trait pallet_prices::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<Self::Event>;
    type Source: DataProvider<CurrencyId, TimeStampedPrice> + DataProviderExtended<CurrencyId, TimeStampedPrice>;
    type FeederOrigin: EnsureOrigin<Self::Origin>;
    type StakingCurrency: Get<CurrencyId>;
    type LiquidCurrency: Get<CurrencyId>;
    type LiquidStakingExchangeRateProvider: ExchangeRateProvider;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<Self::Event>

type Source: DataProvider<CurrencyId, TimeStampedPrice> + DataProviderExtended<CurrencyId, TimeStampedPrice>

The data source, such as Oracle.

type FeederOrigin: EnsureOrigin<Self::Origin>

The origin which may set prices feed to system.

type StakingCurrency: Get<CurrencyId>

Currency used for staking

type LiquidCurrency: Get<CurrencyId>

Currency used for liquid voucher

type LiquidStakingExchangeRateProvider: ExchangeRateProvider

The provider of the exchange rate between liquid currency and staking currency.

Loading content...